maths/clock/clock.go
package clock

import (
	"math"
	"time"
)

const secondHandLength = 90
const clockCentreX = 150
const clockCentreY = 150

// A Point represents a two dimensional Cartesian coordinate
type Point struct {
	X float64
	Y float64
}

// SecondHand is the vector of the second hand of an anologue clock at the given time
func SecondHand(t time.Time) Point {
	p := secondHandPoint(t)
	p = Point{p.X * secondHandLength, p.Y * secondHandLength} // Scale to the length of the hand
	p = Point{p.X, -p.Y}                                      // Flip over X axis to have an origin in the top left hand corner
	p = Point{p.X + clockCentreX, p.Y + clockCentreY}         // Translate to the right position (150,150)

	return p
}

func secondsInRadians(t time.Time) float64 {
	return (math.Pi / (30 / float64(t.Second())))
}

func secondHandPoint(t time.Time) Point {
	angle := secondsInRadians(t)
	x := math.Sin(angle)
	y := math.Cos(angle)

	return Point{x, y}
}
